extract: improve .deb extraction
A Debian package is an 'ar' archive with two files:
- control.tar.gz (metadata)
- data.tar.* (content)

Because cve-bin-tool only cares about the content it can skip iterating the
unpacked archives and go straight to data.tar. This file may be compressed with
a number of compressors so don't hardcode gzip, use glob to find the filename
and let tar determine how to unpack it.

Fixes #2

Signed-off-by: Ross Burton <ross.burton@intel.com>