Prepare docs for ReadTheDocs (#886)
* merged styleguide into contributor guide
* made a bunch of index.rst files to link all the documents
* added doc/_build to gitignore

This is a rough first go to make sure all the docs are at least linked and building. We'll likely want to do some checking to make sure there's no breakages in the Markdown (since it's not the preferred format) and then a lot of refinement so that the document flow works together. But this should be enough to get us started.