test: scanner: Cache downloads and condense them
For test_version_in_package() where we test against a lot of URLs, we
consistently have issues downloading those URLs. Instead of adding all
the files we are downloading to test the scanner on to the git repo,
we've run the recursive extraction and strings and created condensed
versions of the URLs contents.

We will now be able to add these condensed files to our Git repo to
increase speed and stability of testing.

Users adding new tests should commit the added files that appear in
test/condensed-downloads/. Since that is the condensed version of the
URL for the test they've added.

Add test/downloads/ to the .gitignore

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>