Remove CSV file and instead use VendorFetch (#1194)
* Remove pypi_list.csv and use VendorFetch instead

* Use small letters while querying the database

There are packages with names that have capital letters in their names. But their names are written in small letters in the CVE database. So inorder to hit those package values the package name string is in small letters while querying the database.

* Update tests for valid requirements.txt

Append * on the vendor values to show they are guessed vendor values