fix: allow extractraction on all files to fail (#1285)
* fixes #1281
* Related to #1181 

Changes in #1181 to check return codes made it so that cve-bin-tool failed if any file did not extract correctly. 

This changes the default so that failures during extraction are logged but do not halt the scan.  Error messages are not logged when the file extension is .exe because users found those confusing (cve-bin-tool tries to unzip all .exe files in case they are self-extracting zipfiles, but many are not)

Signed-off-by: Terri Oda <terri.oda@intel.com>