fix!: Make error codes fit standard range
Previously, cve-bin-tool returned the number of CVEs found, or negative
numbers for error codes.  Because error codes are expected to be in the
range 0-127 that could cause unexpected behaviours (particularly on
Windows) so this PR changes our error codes to be positive values.

error code 1 is reserved to indicate the cves were found
error codes 2+ indicate operational errors.

Signed-off-by: Terri Oda <terri@toybox.ca>
BREAKING CHANGE: Error codes