ci: disable failing update test on windows (#1963)
The NVD update tests are failing regularly enough on windows in GitHub Actions that it's become a problem for getting new PRs merged.  Since this code is already being run in the linux long tests, I'm disabling it on windows only to see if that allows CI to work more consistently.