fix: fix strings call (#2135)
By default, strings command will only return strings which have at least
4 characters:

"Print sequences of displayable characters that are at least min-len
characters long.  If not specified a default minimum length of 4 is
used."

This is an issue to catch some package versions such as upcoming chrony.
This issue can already be tested by replacing the version of dnsmasq
from 2.80 to 2.8 in mapping_test_data of test/test_data/dnsmasq.py.

To fix this issue, give a minimum length of 3. Indeed, there is very few
packages with a version on one (or two) character(s). The only examples
that can be found on buildroot  are casync and phytool and those
packages don't have any CPEs or CVEs.

Signed-off-by: Fabrice Fontaine <fabrice.fontaine@orange.com>