feat(checker): add libjpeg checker
The version of IJG libjpeg is made of a single digit and a single letter.
This very specific pattern should not raise false positives.

IJG libjpeg should not be confused with
https://github.com/thorfdbg/libjpeg which has the following CPE IDs:
cpe:2.3:a:libjpeg_project:libjpeg and cpe:2.3:a:jpeg:libjpeg

Finally, cve_scanner.py must be updated to handle libjpeg versionning.
To do so, reuse an rename the function already made for openssl

Signed-off-by: Fabrice Fontaine <fabrice.fontaine@orange.com>