ci: skip analyzing things we can't fix in coverity (#2359)
Removes cov-analysis (coverity build directory) and our 3rd party schemas from coverity scans. These were being ignored as "components" in the web interface but were having an effect on our overall defect density and since the cov-analysis directory in particular is huge and filled with errors it was skewing results inaccurately.  This should fix that.