fix: Change extraction exceptions to logged warnings (#2408)
* fixes #1748

Currently, when files can't be extracted because a tool is not installed, we throw an exception and halt the scan.  This is great for getting attention but terrible if you're scanning a large system and maybe just don't care about scanning those files.  It also causes a problem in the issue linked above when attempting to run cve-bin-tool on a system outside of the ones we actually support.

This PR switches us to logging a message and allowing the scan to proceed.