fix: improve product_not_present
If the user adds a valid product in other_products but this product
doesn't have any test in test_data (e.g. open_source for asterisk),
product_not_present.remove(option) will fail.

Add a dedicated message to help the user in this case

While at it, also improve the message when the product is not found in
the test package

Signed-off-by: Fabrice Fontaine <fabrice.fontaine@orange.com>