!fix: generate report when no CVEs detected (#3028) (#3075)
This PR changes fixes an issue where reports were not being generated if no CVEs were found, even if the `--report` flag was used.

This also changes default behaviour of cve-bin-tool so that a report will be generated even if 0 CVEs are found. Previously, reports were only generated when CVEs were found, but as the usage of cve-bin-tool has evolved more people are interested in seeing what components were scanned and using these reports as evidence that no issues were found.  

Users wishing to skip these 0-cve reports and stick to the old behaviour can use the new `--no-0-cve-report` flag to do so. 

Signed-off-by: Pramurta Sinha <contact@b31ngd3v.eu.org>