ci: skip running tests on bot-generated sboms
The earlier PR skipping tests caused issues with our branch protection rules
and had to be disabled because it blocked merging of pull requests (if the
tests were skipped, the PR could never be merged).  This allows part of the job
to run (so it'll pass branch protection checks) while skipping the install and
running of tests on sbom jobs provided by our automated job.

Note that this is the same code as I had in #3446 but I'm separating it out so
it gets a proper code review from someone who is not me.

Signed-off-by: Terri Oda <terri.oda@intel.com>