fix: improve version_compare to drop hashes (#3566)
Currently version_compare raises an error when it gets a hash, since we
can't really compare those without more knowledge of the product. This
changes the code to silently drop hashes and similar distro markers from
the end of versions, which should make our code more robust to real-life
cases.  We may also want to put in an error handler to log the versions
we can't handle without halting the program, but this doesn't do that
yet.

* Adds tests based on #3556 rust versions
* fixes #3556

Signed-off-by: Terri Oda <terri.oda@intel.com>