feat: disable metrics by default (#3618)
Since commit 06b55f7e2aab435e7964fec90ebb30e15ed87585, to get metrics, a
database query is made for every single CVE. This behavior dramatically
increases processing time. For example, 2 seconds was needed before this
change to scan apache2-bin_2.4.10-10+deb8u12_amd64.deb.tar.gz. After
this change, 30 seconds is needed to extract the metrics of the 60 CVEs
related to apache 2.4.10.

Time increase depends on number of CVEs so the increase is even greater
when parsing a directory (e.g. an IoT firmware dump).

I assume that most users are not that interested with those metrics and
would prefer to get a quick result so only retrieve metrics if the user
asks it through the new --metrics option.

Signed-off-by: Fabrice Fontaine <fabrice.fontaine@orange.com>