docs: add cmd for installing the cve-tool in virtualenv
In the contribution guide for running the [local copy of cve-bin-tool](https://github.com/intel/cve-bin-tool/blob/main/CONTRIBUTING.md#running-your-local-copy-of-cve-binary-tool)  ,the mentioned cmd for installing the tool works when a user is not in a virtual env. When present in virtual env , the --user flag isn't needed. 
This PR attempts to clearly mention the cmds for both the cases.