fix: fail gracefully for npm .package-lock.json files (#3654)
* fixes #3559

npm has a .package-lock.json file with a similar name but slightly different
format than the package-lock.json file we can parse with our javascript
language parser.  This changes the javascript parser so it fails more
gracefully if the format of a package-lock.json file does not appear to
be what we can parse.  As in, it prints a warning that it was unable to
parse the file and does not halt the scan.

We should probably build a parser to handle these files correctly in the
future, but for now this will skip them.

Also, I added some docstrings to the files I changed so interrogate
would be happy.

Signed-off-by: Terri Oda <terri.oda@intel.com>