refactor: convert ConfigGenerator class to config_generator function (#3955)
The ConfigGenerator class contains the config_generator method. Currently
config_generator isn't decorated as a class or static method and the
signature doesn't contain 'self'. This has the potential to create issues
in the future if an instance of the ConfigGenerator class is instantiated
or the method requires access to class attributes.

This commit refactors the ConfigGenerator class into a function named
config_generator. A standalone function was chosen over a method
decorator as the ConfigGenerator class maintained no state and contained
a single method. This change removes the potential issues highlighted
above and simplifies the code structure to meet current requirements.