ci: reduce dependabot scan frequency
This changes our dependabot scan frequency to weekly insead of daily in
order to limit the amount of "noise" generated by daily automated PRs.
We could consider switching to monthly if weekly isn't enough reduction.

I've also removed the /doc section in hopes of eliminating the duplicate
PRs we get on doc requirements.

Signed-off-by: Terri Oda <terri.oda@intel.com>