fix: halt if pdf selected but unavailable
* Related to #4326

If pdf output is selected but isn't available, then cve-bin-tool
tells you to install reportlab, but it was easy to miss that message as
the scan continued and pushed it off the screen.  This changes those
messages to be errors insead of info messages, and exits without doing a
scan if pdf is the only output mode selected.

Signed-off-by: Terri Oda <terri.oda@intel.com>